refactor types and call-sites
