+++ code/spikes/publish.001/src/ui/Root.HoldingPattern/entry.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client';
-import { RootHolding } from './index.mjs';
+import { RootHolding } from '.';
 
 /**
  * User Interface
